Rename the lib for foa_energy
