+++ foa_energy_lib.py
